+++ src/XUnity.AutoTranslator.Plugin.Core/Configuration/Settings.cs
@@ -88,6 +88,8 @@ namespace XUnity.AutoTranslator.Plugin.Core.Configuration
       public static TextPostProcessing TranslationPostProcessing;
       public static bool EnableExperimentalHooks;
       public static bool ForceExperimentalHooks;
+      public static bool CacheRegexLookups;
+      public static bool CacheWhitespaceDifferences;
 
       public static string TextureDirectory;
       public static bool EnableTextureTranslation;
@@ -146,6 +148,8 @@ namespace XUnity.AutoTranslator.Plugin.Core.Configuration
             TranslationPostProcessing = PluginEnvironment.Current.Preferences.GetOrDefault( "Behaviour", "TranslationPostProcessing", TextPostProcessing.ReplaceMacronWithCircumflex );
             EnableExperimentalHooks = PluginEnvironment.Current.Preferences.GetOrDefault( "Behaviour", "EnableExperimentalHooks", false );
             ForceExperimentalHooks = PluginEnvironment.Current.Preferences.GetOrDefault( "Behaviour", "ForceExperimentalHooks", false );
+            CacheRegexLookups = PluginEnvironment.Current.Preferences.GetOrDefault( "Behaviour", "CacheRegexLookups", true );
+            CacheWhitespaceDifferences = PluginEnvironment.Current.Preferences.GetOrDefault( "Behaviour", "CacheWhitespaceDifferences", true );
 
             TextureDirectory = PluginEnvironment.Current.Preferences.GetOrDefault( "Texture", "TextureDirectory", @"Translation\Texture" );
             EnableTextureTranslation = PluginEnvironment.Current.Preferences.GetOrDefault( "Texture", "EnableTextureTranslation", false );